% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-checkmate.R
\name{chk_choice}
\alias{chk_choice}
\title{Modify how \code{checkmate} handles output of \code{checkmate::check_choice}}
\usage{
chk_choice(..., quiet = TRUE)
}
\arguments{
\item{...}{Arguments passed on to the specified \code{checkmate} function.}

\item{quiet}{A logical which is TRUE if a failed check should not emit its
error message; FALSE if it should. Default: TRUE}
}
\value{
A logical which is TRUE if the check succeeded and FALSE if not. The
message generated by a failed check is set as an attribute of the return
value.
}
\description{
Modify how \code{checkmate} handles output of \code{checkmate::check_choice}
}
