% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-cli.R
\name{glimpse_arg}
\alias{glimpse_arg}
\title{Format a multiline argument for printing}
\usage{
glimpse_arg(arg, max_lines = 3L, bullet_theme = " ")
}
\arguments{
\item{arg}{The argument to print}

\item{max_lines}{The maximum number of lines of output to show for the
printed argument. This would include, for example, a line for the column
numbers/names of a matrix (e.g. "[,1]  [,2]").}

\item{bullet_theme}{A character for the \code{cli::cli_bullets} theme.}
}
\value{
A named character vector which is the result of printing the
argument, up to a maximum of \code{max_lines}, named by the value of
\code{bullet_theme}.
}
\description{
Formats a multiline argument for printing, retaining only a given number of
lines to display
}
\examples{
x <- matrix(c(6.62,      4.9,     1.94,
              4.9,      -3.524,  10.5,
              100.94, -678.5,   -20.95), 3, 3)
cli::cli_verbatim(glimpse_arg(x))
# Alternatively:
cli::cli_inform(cli::cat_line(glimpse_arg(x)))

}
\seealso{
\code{\link[cli:cli_verbatim]{cli::cli_verbatim()}}
}
