% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-checkmate.R
\name{chk_choice}
\alias{chk_choice}
\title{Wrap \code{checkmate::check_choice} to handle its output differently}
\usage{
chk_choice(..., quiet = TRUE)
}
\arguments{
\item{...}{Arguments passed on to the specified \code{checkmate} function.}

\item{quiet}{A logical which is TRUE if a failed check should not emit its
error message; FALSE if it should. Default: TRUE}
}
\value{
A logical which is TRUE if the check succeeded and FALSE if not. The
message generated by a failed check is set as an attribute of the return
value.
}
\description{
This wrapper always returns TRUE or FALSE so that it can be used in
conditional statements. If the check fails, it will optionally emit the error
message returned by checkmate.
}
