% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simssd-package.R
\docType{package}
\name{simssd-package}
\alias{simssd}
\alias{simssd-package}
\title{simssd: Faster Simulation for Multilevel Sample Size Determination (SSD)}
\description{
Performs sample size determination (SSD) and power computation via simulation for fixed effects in multilevel linear regression models with an emphasis on improving computational speed.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/toniprice/simssd}
  \item \url{https://toniprice.github.io/simssd/}
  \item Report bugs at \url{https://github.com/toniprice/simssd/issues}
}

}
\author{
\strong{Maintainer}: Toni Price \email{placeholder@example.com} (\href{https://orcid.org/0000-0002-9109-5011}{ORCID})

}
\keyword{internal}
